+++ src/main/java/roomescape/controller/console/ReservationConsoleController.java
@@ -1,9 +1,9 @@
 package roomescape.controller.console;
 
 import org.springframework.stereotype.Controller;
-import roomescape.dto.ReservationRequest;
-import roomescape.dto.ReservationResponse;
-import roomescape.dto.ReservationTimeResponse;
+import roomescape.dto.request.ReservationRequest;
+import roomescape.dto.response.ReservationResponse;
+import roomescape.dto.response.ReservationTimeResponse;
 import roomescape.service.ReservationService;
 import roomescape.service.ReservationTimeService;
 import roomescape.view.ReservationTimeView;
