+++ src/main/java/roomescape/controller/web/ReservationController.java
@@ -2,9 +2,9 @@ package roomescape.controller.web;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import roomescape.dto.ReservationRequest;
-import roomescape.dto.ReservationResponse;
-import roomescape.dto.SelectableTimeResponse;
+import roomescape.dto.request.ReservationRequest;
+import roomescape.dto.response.ReservationResponse;
+import roomescape.dto.response.SelectableTimeResponse;
 import roomescape.service.ReservationService;
 
 import java.net.URI;
