--- src/main/java/roomescape/dto/SelectableTimeRequest.java
@@ -1,9 +1,0 @@
-package roomescape.dto;
-
-import java.time.LocalDate;
-
-public record SelectableTimeRequest(
-        LocalDate date,
-        long themeId
-) {
-}
