+++ src/main/java/roomescape/dto/request/ReservationRequest.java
@@ -1,9 +1,10 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
-import java.time.LocalDate;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
+
+import java.time.LocalDate;
 
 public record ReservationRequest(
         String name,
