+++ src/main/java/roomescape/dto/response/SelectableTimeResponse.java
@@ -1,0 +1,15 @@
+package roomescape.dto.response;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+
+import java.time.LocalTime;
+
+public record SelectableTimeResponse(
+        long timeId,
+
+        @JsonFormat(pattern = "kk:mm")
+        LocalTime startAt,
+
+        boolean alreadyBooked
+) {
+}
